Add table_of_functions template; start using it
In order to avoid having to maintain the tables of functions every time
we add a new function, I decided to do it programmatically using the new
table_of_functions template.

In this commit I demonstrate how to use it in
h2spatial-ext/affine-transformations. The following commits will move
the descriptions to the corresponding files.