Fix unit test for U
Important: See the proof of Theorem 6 in Brande's algorithm. He supposes
that on any shortest path from s to t containing v there is *exactly
one* edge {v,w} with v in P_s(w).

This is obviously false if there are multiple edges with the same
weight, as is the case in this example (U).

TODO? Adapt Brande's algorithm to this case.