Add test for centrality of a disconnected graph.
If the graph is not connected (or not strongly connected, for directed
graphs), all closeness values will be zero, since for any node i, there
exists a node j with d(i,j)=Infinity. Hence
    C_C(i) = (N-1)/(... + Infinity + ...) = 0
But i was arbitrary.

Both node and edge betweenness values, on the otherhand, will not
necessarily be zero.