Allow no geometry column in input table
Before, we always required the input table to have a geometry column for
ST_ShortestPath and ST_ShortestPathTree. This was so that when using
these functions with a GIS, we wouldn't have to do a JOIN before
visualizing the results.

Now, the output table will have a geometry column if and only if the
input table does. This keeps backward compatibility while allowing
calculations on graphs whose edges do not have corresponding geometries.

Now any user working with edge geometries has the freedom of deciding
when he wants to do the JOIN:

* If done before (so that the input table has a geometry column), all
invocations of either function will produce results having geometry
columns. This is the same as the old behavior and saves time.
* It is still possible to do the join after.